DTFS2-4177 : route renamed, test added, lint fixes
