+++ gef-ui/server/routes/__tests__/confirm-cover-start-date.test.js
@@ -22,6 +22,7 @@ describe('Routes', () => {
   });
 
   it('Sets up all routes', () => {
-    expect(getSpy).toHaveBeenCalledWith('/application-details/:applicationId/confirm-cover-start-date', validateToken, expect.any(Function));
+    expect(getSpy).toHaveBeenCalledWith('/application-details/:applicationId/:facilityId/confirm-cover-start-date/', validateToken, expect.any(Function));
+    expect(postSpy).toHaveBeenCalledWith('/application-details/:applicationId/:facilityId/confirm-cover-start-date/', validateToken, expect.any(Function));
   });
 });
