--- gef-ui/server/routes/ukef-cover-start-date.js
@@ -1,11 +1,0 @@
-const express = require('express');
-const { processCoverStartDate } = require('../controllers/ukef-cover-start-date');
-const { applicationDetails } = require('../controllers/application-details');
-const validateToken = require('../middleware/validateToken');
-
-const router = express.Router();
-
-router.get('/application-details/:applicationId/:facilityId/ukef-cover-start-date/', validateToken, (req, res) => applicationDetails(req, res));
-router.post('/application-details/:applicationId/:facilityId/ukef-cover-start-date/', validateToken, (req, res) => processCoverStartDate(req, res));
-
-module.exports = router;
